Fix x86 warnings threated as errors.
--HG--
branch : PointoolsVortexAPI

+++ iModelBridgeCore/Pointools/PointoolsVortexAPI/src/ptclash/clashSet.cpp
@@ -28,7 +28,7 @@ void ClashSet::addObject( pt::Object3D *obj )
 //-----------------------------------------------------------------------------
 const ClashObject* ClashSet::hasObject( const pt::PersistentObjectRef &obj ) const
 {
-	for (int i=0; i<m_objects.size(); i++)
+	for (size_t i=0; i<m_objects.size(); i++)
 	{
 		const ClashObject *co = m_objects[i];
 		if (co->objRef() == obj) return co;
@@ -38,7 +38,7 @@ const ClashObject* ClashSet::hasObject( const pt::PersistentObjectRef &obj ) con
 //-----------------------------------------------------------------------------
 ClashObject* ClashSet::getObject( const pt::PersistentObjectRef &obj ) const
 {
-	for (int i=0; i<m_objects.size(); i++)
+	for (size_t i=0; i<m_objects.size(); i++)
 	{
 		ClashObject *co = m_objects[i];
 		if (co->objRef() == obj) return co;
@@ -48,7 +48,7 @@ ClashObject* ClashSet::getObject( const pt::PersistentObjectRef &obj ) const
 //-----------------------------------------------------------------------------
 bool ClashSet::hasObject( const ClashObject* obj ) const
 {
-	for (int i=0; i<m_objects.size(); i++)
+	for (size_t i=0; i<m_objects.size(); i++)
 	{
 		if (obj == m_objects[i]) return true;		
 	}
@@ -66,7 +66,7 @@ void ClashSet::addObject( const pt::PersistentObjectRef &obj )
 //-----------------------------------------------------------------------------
 int  ClashSet::getObjects( std::vector<pt::PersistentObjectRef> &objs ) const
 {
-	for (unsigned int i=0; i<m_objects.size(); i++)
+	for (size_t i=0; i<m_objects.size(); i++)
 	{
 		objs.push_back( m_objects[i]->objRef() );
 	}
@@ -75,7 +75,7 @@ int  ClashSet::getObjects( std::vector<pt::PersistentObjectRef> &objs ) const
 //-----------------------------------------------------------------------------
 void ClashSet::clear()
 {
-	for (unsigned int i=0; i<m_objects.size(); i++)
+	for (size_t i=0; i<m_objects.size(); i++)
 	{
 		delete m_objects[i];
 	}
@@ -86,7 +86,7 @@ void ClashSet::remObject( const pt::PersistentObjectRef &obj )
 {
 	std::vector<ClashObject*> objs;
 
-	for (unsigned int i=0; i<m_objects.size(); i++)
+	for (size_t i=0; i<m_objects.size(); i++)
 	{
 		if (m_objects[i]->objRef() == obj)
 		{
@@ -104,7 +104,7 @@ void ClashSet::remObject( ClashObject *obj )
 {
 	std::vector<ClashObject*> objs;
 
-	for (unsigned int i=0; i<m_objects.size(); i++)
+	for (size_t i=0; i<m_objects.size(); i++)
 	{
 		if (m_objects[i] == obj)
 		{
@@ -123,7 +123,7 @@ int ClashSet::drawTrees(int depth)
 {
 	int count = 0;
 
-	for (unsigned int i=0; i<m_objects.size(); i++)
+	for (size_t i=0; i<m_objects.size(); i++)
 	{
 		if (m_objects[i])
 			count += m_objects[i]->drawObjBoundsTree(depth);
@@ -137,7 +137,7 @@ void ClashSet::getDiagnosticString( pt::String &str ) const
 	int max_depth = 0;
 	int min_depth = 200;
 
-	for (unsigned int i=0; i<m_objects.size(); i++)
+	for (size_t i=0; i<m_objects.size(); i++)
 	{
 		if (m_objects[i])
 		{
